requirements.txt
fastapi>=0.100.0
uvicorn[standard]
pyjwt
passlib[bcrypt]
python-multipart
pydantic[email]
redis
httpx
pytest
pytest-asyncio
pytest-cov
python-dotenv
